Use `obfuscate` utility instead of hardcoded "*" * 8
