+++ tests/orion/models/test_block_documents.py
@@ -9,7 +9,7 @@ from pydantic import SecretBytes, SecretStr
 from prefect.blocks.core import Block
 from prefect.orion import models, schemas
 from prefect.orion.schemas.actions import BlockDocumentCreate
-from prefect.orion.utilities.names import obfuscate_string
+from prefect.utilities.names import obfuscate_string
 
 
 def long_string(s: str):
