+++ tests/utilities/test_names.py
@@ -1,6 +1,6 @@
 import pytest
 
-from prefect.orion.utilities.names import obfuscate_string
+from prefect.utilities.names import obfuscate, obfuscate_string
 
 
 @pytest.mark.parametrize(
@@ -24,3 +24,21 @@ def test_obfuscate_string(s, expected):
     assert obfuscate_string(s) == obfuscate_string(s, show_tail=False) == "*" * 8
     # show tail
     assert obfuscate_string(s, show_tail=True) == expected
+
+
+@pytest.mark.parametrize(
+    "s, expected",
+    [
+        (None, "********"),
+        ("", "********"),
+        ("a", "********"),
+        ("abcdefghijklm", "*****klm"),
+        (1, "********"),
+        ({"x": "y"}, "********"),
+    ],
+)
+def test_obfuscate(s, expected):
+    # default is not to reveal any characters
+    assert obfuscate(s) == obfuscate(s, show_tail=False) == "*" * 8
+    # show tail
+    assert obfuscate(s, show_tail=True) == expected
